# contracts/openapi.json
{
  "components": {
    "schemas": {
      "CameraCreateRequest": {
        "additionalProperties": false,
        "description": "创建完整 Camera 聚合的请求。",
        "examples": [
          {
            "ip_address": "192.0.2.64",
            "name": "洗手区 01",
            "password": "openapi-test-password",
            "rtsp_port": 554,
            "sources": [
              {
                "is_default_preview": true,
                "name": "主码流",
                "url_suffix": "Streaming/Channels/101"
              },
              {
                "is_default_preview": false,
                "name": "子码流",
                "url_suffix": "/Streaming/Channels/102"
              }
            ],
            "username": "openapi-test-user"
          }
        ],
        "properties": {
          "ip_address": {
            "format": "ipv4",
            "title": "Ip Address",
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "password": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "rtsp_port": {
            "default": 554,
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Rtsp Port",
            "type": "integer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/CameraSourceCreateRequest"
            },
            "minItems": 1,
            "title": "Sources",
            "type": "array"
          },
          "username": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "name",
          "ip_address",
          "username",
          "password",
          "sources"
        ],
        "title": "CameraCreateRequest",
        "type": "object"
      },
...
      "CameraPage": {
        "additionalProperties": false,
        "description": "稳定分页的 Camera 列表响应。",
        "examples": [
          {
            "items": [
              {
                "camera_id": "6f9619ff-8b86-4e4f-9f68-bb3f8f6f4f21",
                "created_at": "2026-08-19T03:00:00Z",
                "default_preview_source": {
                  "last_checked_at": "2026-08-19T03:10:01Z",
                  "name": "主码流",
                  "source_id": "3f2504e0-4f89-41d3-9a0c-0305e82c3301",
                  "status": "ONLINE",
                  "whep_url": "https://media.example.invalid/3f2504e0-4f89-41d3-9a0c-0305e82c3301/whep"
                },
                "ip_address": "192.0.2.64",
                "name": "洗手区 01",
                "online_source_count": 2,
                "rtsp_port": 554,
                "source_count": 2,
                "status": "ONLINE",
                "updated_at": "2026-08-19T03:10:00Z"
              }
            ],
            "page": 1,
            "page_size": 20,
            "total": 1
          }
        ],
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CameraSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "CameraPage",
        "type": "object"
      },
...
      "JsonValue": {},
      "PlaybackInfo": {
        "additionalProperties": false,
        "description": "不含上游连接信息的 WHEP 播放准备结果。",
        "examples": [
          {
            "protocol": "WHEP",
            "source_id": "8f14e45f-ea9d-4a7d-9b6d-8c9f0a1b2c3d",
            "status": "AVAILABLE",
            "url": "https://media.example.invalid/8f14e45f-ea9d-4a7d-9b6d-8c9f0a1b2c3d/whep"
          }
        ],
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "protocol": {
            "$ref": "#/components/schemas/PlaybackProtocol"
          },
          "source_id": {
            "format": "uuid4",
            "title": "Source Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PlaybackStatus"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "protocol",
          "url",
          "status",
          "expires_at"
        ],
        "title": "PlaybackInfo",
        "type": "object"
      },
      "PlaybackProtocol": {
        "description": "MVP 唯一支持的浏览器播放协议。",
        "enum": [
          "WHEP"
        ],
        "title": "PlaybackProtocol",
        "type": "string"
      },
      "PlaybackStatus": {
        "description": "只有媒体 Path 严格在线时才返回的成功状态。",
        "enum": [
          "AVAILABLE"
        ],
        "title": "PlaybackStatus",
        "type": "string"
      },
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "所有结构化 HTTP 错误共享的唯一响应模型。\n\n``context`` 只允许 JSON 值；调用工厂的映射代码仍需只放入已批准的非敏感字段，不能把\n原始异常、请求体或第三方响应整体塞入该字典。冻结和禁止额外字段也确保步骤 6 注册到\nOpenAPI 的模型与实际运行时响应保持同源。",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "context": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Context",
            "type": "object"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "instance",
          "trace_id"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "SetDefaultPreviewSourceRequest": {
        "additionalProperties": false,
        "description": "把一个属于当前 Camera 的 Source 设为默认预览源。",
        "examples": [
          {
            "source_id": "3f2504e0-4f89-41d3-9a0c-0305e82c3301"
          }
        ],
        "properties": {
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          }
        },
        "required": [
          "source_id"
        ],
        "title": "SetDefaultPreviewSourceRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SOP Vision 后端",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/camera-sources/{source_id}/playback": {
      "get": {
        "operationId": "getCameraSourcePlayback",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "protocol": "WHEP",
                  "source_id": "8f14e45f-ea9d-4a7d-9b6d-8c9f0a1b2c3d",
                  "status": "AVAILABLE",
                  "url": "https://media.example.invalid/8f14e45f-ea9d-4a7d-9b6d-8c9f0a1b2c3d/whep"
                },
                "schema": {
                  "$ref": "#/components/schemas/PlaybackInfo"
                }
              }
            },
            "description": "返回已就绪的 WHEP 播放地址。",
            "headers": {
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "code": "NOT_FOUND",
                  "context": {},
                  "detail": "请求未能完成。",
                  "errors": [],
                  "instance": "/api/v1/cameras",
                  "status": 404,
                  "title": "资源不存在",
                  "trace_id": "tr_openapi_example",
                  "type": "urn:sop-vision:problem:not-found"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "HTTP 404 Problem Details",
            "headers": {
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/problem+json": {
                "example": {
                  "code": "PLAYBACK_NOT_AVAILABLE",
                  "context": {},
                  "detail": "请求未能完成。",
                  "errors": [],
                  "instance": "/api/v1/cameras",
                  "status": 409,
                  "title": "播放尚不可用",
                  "trace_id": "tr_openapi_example",
                  "type": "urn:sop-vision:problem:playback-not-available"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "HTTP 409 Problem Details",
            "headers": {
              "Retry-After": {
                "description": "建议客户端再次检查播放可用性的秒数。",
                "schema": {
                  "minimum": 0,
                  "type": "integer"
                }
              },
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "example": {
                  "code": "VALIDATION_ERROR",
                  "context": {},
                  "detail": "请求未能完成。",
                  "errors": [
                    {
                      "code": "REQUIRED",
                      "detail": "该字段为必填项。",
                      "field": "sources[0].name"
                    }
                  ],
                  "instance": "/api/v1/cameras",
                  "status": 422,
                  "title": "请求字段验证失败",
                  "trace_id": "tr_openapi_example",
                  "type": "urn:sop-vision:problem:validation-error"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "HTTP 422 Problem Details",
            "headers": {
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "502": {
            "content": {
              "application/problem+json": {
                "example": {
                  "code": "MEDIA_SERVICE_INVALID_RESPONSE",
                  "context": {},
                  "detail": "请求未能完成。",
                  "errors": [],
                  "instance": "/api/v1/cameras",
                  "status": 502,
                  "title": "媒体服务响应无效",
                  "trace_id": "tr_openapi_example",
                  "type": "urn:sop-vision:problem:media-service-invalid-response"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "HTTP 502 Problem Details",
            "headers": {
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "example": {
                  "code": "SERVICE_UNAVAILABLE",
                  "context": {},
                  "detail": "请求未能完成。",
                  "errors": [],
                  "instance": "/api/v1/cameras",
                  "status": 503,
                  "title": "服务暂不可用",
                  "trace_id": "tr_openapi_example",
                  "type": "urn:sop-vision:problem:service-unavailable"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "HTTP 503 Problem Details",
            "headers": {
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Get Camera Source Playback",
        "tags": [
          "camera-sources"
        ]
      }
    },
    "/api/v1/cameras": {
      "get": {
        "operationId": "listCameras",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "items": [
                    {
                      "camera_id": "6f9619ff-8b86-4e4f-9f68-bb3f8f6f4f21",
                      "created_at": "2026-08-19T03:00:00Z",
                      "default_preview_source": {
                        "last_checked_at": "2026-08-19T03:10:01Z",
                        "name": "主码流",
                        "source_id": "3f2504e0-4f89-41d3-9a0c-0305e82c3301",
                        "status": "ONLINE",
                        "whep_url": "https://media.example.invalid/3f2504e0-4f89-41d3-9a0c-0305e82c3301/whep"
                      },
                      "ip_address": "192.0.2.64",
                      "name": "洗手区 01",
                      "online_source_count": 2,
                      "rtsp_port": 554,
                      "source_count": 2,
                      "status": "ONLINE",
                      "updated_at": "2026-08-19T03:10:00Z"
                    }
                  ],
                  "page": 1,
                  "page_size": 20,
                  "total": 1
                },
                "schema": {
                  "$ref": "#/components/schemas/CameraPage"
                }
              }
            },
            "description": "返回一页 Camera 摘要。",
            "headers": {
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "example": {
                  "code": "VALIDATION_ERROR",
                  "context": {},
                  "detail": "请求未能完成。",
                  "errors": [
                    {
                      "code": "REQUIRED",
                      "detail": "该字段为必填项。",
                      "field": "sources[0].name"
                    }
                  ],
                  "instance": "/api/v1/cameras",
                  "status": 422,
                  "title": "请求字段验证失败",
                  "trace_id": "tr_openapi_example",
                  "type": "urn:sop-vision:problem:validation-error"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "HTTP 422 Problem Details",
            "headers": {
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "example": {
                  "code": "SERVICE_UNAVAILABLE",
                  "context": {},
                  "detail": "请求未能完成。",
                  "errors": [],
                  "instance": "/api/v1/cameras",
                  "status": 503,
                  "title": "服务暂不可用",
                  "trace_id": "tr_openapi_example",
                  "type": "urn:sop-vision:problem:service-unavailable"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "HTTP 503 Problem Details",
            "headers": {
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "List Cameras",
        "tags": [
          "cameras"
        ]
      },
      "post": {
        "operationId": "createCamera",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CameraCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "camera_id": "6f9619ff-8b86-4e4f-9f68-bb3f8f6f4f21",
                  "created_at": "2026-08-19T03:00:00Z",
                  "default_preview_source_id": "3f2504e0-4f89-41d3-9a0c-0305e82c3301",
                  "ip_address": "192.0.2.64",
                  "name": "洗手区 01",
                  "online_source_count": 1,
                  "password": "openapi-test-password",
                  "rtsp_port": 554,
                  "source_count": 2,
                  "sources": [
                    {
                      "is_default_preview": true,
                      "last_checked_at": "2026-08-19T03:10:01Z",
                      "name": "主码流",
                      "rtsp_url": "rtsp://openapi-test-user:openapi-test-password@192.0.2.64:554/Streaming/Channels/101",
                      "source_id": "3f2504e0-4f89-41d3-9a0c-0305e82c3301",
                      "status": "ONLINE",
                      "url_suffix": "Streaming/Channels/101",
                      "whep_url": "https://media.example.invalid/3f2504e0-4f89-41d3-9a0c-0305e82c3301/whep"
                    },
                    {
                      "error": "MTX_PATH_NOT_FOUND",
                      "is_default_preview": false,
                      "last_checked_at": "2026-08-19T03:10:01Z",
                      "name": "子码流",
                      "rtsp_url": "rtsp://openapi-test-user:openapi-test-password@192.0.2.64:554/Streaming/Channels/102",
                      "source_id": "8f14e45f-ea9d-4a7d-9b6d-8c9f0a1b2c3d",
                      "status": "OFFLINE",
                      "url_suffix": "Streaming/Channels/102"
                    }
                  ],
                  "status": "DEGRADED",
                  "updated_at": "2026-08-19T03:10:00Z",
                  "username": "openapi-test-user"
                },
                "schema": {
                  "$ref": "#/components/schemas/CameraDetail"
                }
              }
            },
            "description": "Camera 聚合创建成功。",
            "headers": {
              "Cache-Control": {
                "description": "敏感 CameraDetail 不得被共享或持久化缓存。",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "Location": {
                "description": "新建 Camera 的规范详情路径。",
                "schema": {
                  "type": "string"
                }
              },
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "example": {
                  "code": "VALIDATION_ERROR",
                  "context": {},
                  "detail": "请求未能完成。",
                  "errors": [
                    {
                      "code": "REQUIRED",
                      "detail": "该字段为必填项。",
                      "field": "sources[0].name"
                    }
                  ],
                  "instance": "/api/v1/cameras",
                  "status": 422,
                  "title": "请求字段验证失败",
                  "trace_id": "tr_openapi_example",
                  "type": "urn:sop-vision:problem:validation-error"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "HTTP 422 Problem Details",
            "headers": {
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "example": {
                  "code": "SERVICE_UNAVAILABLE",
                  "context": {},
                  "detail": "请求未能完成。",
                  "errors": [],
                  "instance": "/api/v1/cameras",
                  "status": 503,
                  "title": "服务暂不可用",
                  "trace_id": "tr_openapi_example",
                  "type": "urn:sop-vision:problem:service-unavailable"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "HTTP 503 Problem Details",
            "headers": {
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Create Camera",
        "tags": [
          "cameras"
        ]
      }
    },
...
    "/api/v1/cameras/{camera_id}/default-preview-source": {
      "patch": {
        "operationId": "setDefaultPreviewSource",
        "parameters": [
          {
            "in": "path",
            "name": "camera_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Camera Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDefaultPreviewSourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "camera_id": "6f9619ff-8b86-4e4f-9f68-bb3f8f6f4f21",
                  "default_preview_source_id": "3f2504e0-4f89-41d3-9a0c-0305e82c3301",
                  "updated_at": "2026-08-19T03:10:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/DefaultPreviewSourceResponse"
                }
              }
            },
            "description": "默认预览 Source 切换成功。",
            "headers": {
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "example": {
                  "code": "NOT_FOUND",
                  "context": {},
                  "detail": "请求未能完成。",
                  "errors": [],
                  "instance": "/api/v1/cameras",
                  "status": 404,
                  "title": "资源不存在",
                  "trace_id": "tr_openapi_example",
                  "type": "urn:sop-vision:problem:not-found"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "HTTP 404 Problem Details",
            "headers": {
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "example": {
                  "code": "VALIDATION_ERROR",
                  "context": {},
                  "detail": "请求未能完成。",
                  "errors": [
                    {
                      "code": "REQUIRED",
                      "detail": "该字段为必填项。",
                      "field": "sources[0].name"
                    }
                  ],
                  "instance": "/api/v1/cameras",
                  "status": 422,
                  "title": "请求字段验证失败",
                  "trace_id": "tr_openapi_example",
                  "type": "urn:sop-vision:problem:validation-error"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "HTTP 422 Problem Details",
            "headers": {
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "example": {
                  "code": "SERVICE_UNAVAILABLE",
                  "context": {},
                  "detail": "请求未能完成。",
                  "errors": [],
                  "instance": "/api/v1/cameras",
                  "status": 503,
                  "title": "服务暂不可用",
                  "trace_id": "tr_openapi_example",
                  "type": "urn:sop-vision:problem:service-unavailable"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "HTTP 503 Problem Details",
            "headers": {
              "X-Trace-Id": {
                "description": "本次请求的关联 ID；与 Problem body 的 trace_id 相同。",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Set Default Preview Source",
        "tags": [
          "cameras"
        ]
      }
    },
...
  }
}